refactor(presentation): clean architecture
