--- app/src/main/java/com/kristallik/jokeapp/ui/add_joke/AddJokeView.kt
@@ -1,7 +1,0 @@
-package com.kristallik.jokeapp.ui.add_joke
-
-interface AddJokeView {
-    fun showCategories(categories: List<String>)
-    fun showError(errorMessage: String)
-    fun saveJoke(message: String)
-}
