--- app/src/main/java/com/kristallik/jokeapp/ui/main/MainActivity.kt
@@ -1,23 +1,0 @@
-package com.kristallik.jokeapp.ui.main
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import androidx.fragment.app.commit
-import com.kristallik.jokeapp.R
-import com.kristallik.jokeapp.databinding.ActivityMainBinding
-import com.kristallik.jokeapp.ui.main.fragment.JokeListFragment
-
-class MainActivity : AppCompatActivity() {
-    private lateinit var binding: ActivityMainBinding
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = ActivityMainBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-        if (savedInstanceState == null) {
-            supportFragmentManager.commit {
-                replace(R.id.fragment_container, JokeListFragment())
-            }
-        }
-    }
-}
