--- app/src/main/java/com/kristallik/jokeapp/ui/main/MainView.kt
@@ -1,10 +1,0 @@
-package com.kristallik.jokeapp.ui.main
-
-import com.kristallik.jokeapp.data.Joke
-
-interface MainView {
-    suspend fun showJokes(jokes: List<Joke>)
-    fun showError(errorMessage: String)
-    fun addJoke()
-    fun showToast(message: String)
-}
